MAJOR Update - GUI Changes, Function Changes, Added Functions, Added Image Assets
The app UI has been overhauled to better reflect how it looks on the
ingame HUD by changing the color schemes.

The app now has a launch ICON that I created.

The REFILL button has been changed to RESET. Instead of simply refilling
the ammo, it will reset the sentry variables (health, ammo) and the PDA
UI.

Upon the start of the bluetooth connection, the SAP and S+F buttons will
be disabled as they will now be treated as debug buttons only.

Added a new image asset. The wrench icon will appear when the sentry has
reached a certain health or reached a certain amount of remaining ammo.

The sapper icon asset has been deleted due to it's poor quality, due to
being the wrong kind of assset. It's been replaced as a drawable image
resource which fixes the quality issue.

Some functions have been renamed to better pinpoint what they are for,
and some new functions were added, for example, a separate animation for
the wrench icon (it is 2x slower than the sapper icon animation). Others
were modified. For example, the mReceiver has a new else if to receive
from the STM that the sentry was unsapped.

Moved some functions and added some comments to make the code easier to
read.